% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseQ.R
\name{parseQ}
\alias{parseQ}
\title{Parse a quadratic equation into a matrix of coefficients}
\usage{
parseQ(model, q)
}
\arguments{
\item{model}{An Rglpk model object}

\item{q}{A string containting the quadratic portion of a CPLEX constraint}
}
\value{
A long matrix of variable pairs and their coefficients
}
\description{
Take a glpk model object a Q equation and return a long-form DF of variable/coef pairs
}

