% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQP.R
\name{readQP}
\alias{readQP}
\title{Utility for reading model files for parsing}
\usage{
readQP(filePath)
}
\arguments{
\item{filePath}{The path to the model file}
}
\value{
A character string containing the contents of the model file.
}
\description{
Read a CPLEX model file into memory as a character string
}

